+++ frontend/src/components/ui/index.ts
@@ -1,5 +1,6 @@
 export { default as Button } from './Button'
 export { Card, CardHeader, CardTitle, CardContent } from './Card'
+export { Form, FormField, FormSection } from './Form'
 export { default as Input } from './Input'
 export { default as Modal } from './Modal'
 export { default as Select } from './Select'
